Move the build script into a subdir
