+++ Cargo.toml
@@ -6,12 +6,16 @@ authors = [
   "Jake Goulding <jake.goulding@gmail.com>",
   "Dylan McKay <me@dylanmckay.io>",
 ]
+
 license = "MIT/Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/avr-rust/ruduino"
 description = """
 Reusable components for the Arduino Uno.
 """
+
+build = "core_generator/build.rs"
+
 keywords = ["avr", "arduino", "uno"]
 
 [build-dependencies]
